Moved interfaces back to where the service classes were initially.
