+++ src/main/java/com/mangopay/core/APIs/ApiCards.java
@@ -1,4 +1,4 @@
-package com.mangopay.core.APIs.declaration;
+package com.mangopay.core.APIs;
 
 import com.mangopay.entities.Card;
 import com.mangopay.entities.TemporaryPaymentCard;
