+++ src/main/java/com/mangopay/core/APIs/ApiKycDocuments.java
@@ -1,4 +1,4 @@
-package com.mangopay.core.APIs.declaration;
+package com.mangopay.core.APIs;
 
 import com.mangopay.core.Pagination;
 import com.mangopay.core.Sorting;
