+++ src/main/java/com/mangopay/core/APIs/ApiPayIns.java
@@ -1,4 +1,4 @@
-package com.mangopay.core.APIs.declaration;
+package com.mangopay.core.APIs;
 
 import com.mangopay.entities.PayIn;
 import com.mangopay.entities.Refund;
