+++ src/main/java/com/mangopay/core/APIs/implementation/ApiCardRegistrationsImpl.java
@@ -2,7 +2,7 @@ package com.mangopay.core.APIs.implementation;
 
 import com.mangopay.MangoPayApi;
 import com.mangopay.core.APIs.ApiBase;
-import com.mangopay.core.APIs.declaration.ApiCardRegistrations;
+import com.mangopay.core.APIs.ApiCardRegistrations;
 import com.mangopay.entities.CardRegistration;
 
 /**
