+++ src/main/java/com/mangopay/core/APIs/implementation/ApiDisputesImpl.java
@@ -3,11 +3,10 @@ package com.mangopay.core.APIs.implementation;
 import com.mangopay.MangoPayApi;
 import com.mangopay.core.*;
 import com.mangopay.core.APIs.ApiBase;
-import com.mangopay.core.APIs.declaration.ApiDisputes;
+import com.mangopay.core.APIs.ApiDisputes;
 import com.mangopay.entities.Dispute;
 import com.mangopay.entities.DisputeDocument;
 import com.mangopay.entities.DisputePage;
-import com.mangopay.entities.KycPage;
 import com.mangopay.entities.Repudiation;
 import com.mangopay.entities.SettlementTransfer;
 import com.mangopay.entities.Transaction;
