+++ src/main/java/com/mangopay/core/APIs/implementation/ApiHooksImpl.java
@@ -3,7 +3,7 @@ package com.mangopay.core.APIs.implementation;
 import com.mangopay.MangoPayApi;
 import com.mangopay.core.*;
 import com.mangopay.core.APIs.ApiBase;
-import com.mangopay.core.APIs.declaration.ApiHooks;
+import com.mangopay.core.APIs.ApiHooks;
 import com.mangopay.entities.Hook;
 import java.util.List;
