+++ src/main/java/com/mangopay/core/APIs/implementation/ApiPayOutsImpl.java
@@ -2,7 +2,7 @@ package com.mangopay.core.APIs.implementation;
 
 import com.mangopay.MangoPayApi;
 import com.mangopay.core.APIs.ApiBase;
-import com.mangopay.core.APIs.declaration.ApiPayOuts;
+import com.mangopay.core.APIs.ApiPayOuts;
 import com.mangopay.entities.PayOut;
 
 /**
