+++ src/main/java/com/mangopay/core/APIs/implementation/ApiRefundsImpl.java
@@ -2,7 +2,7 @@ package com.mangopay.core.APIs.implementation;
 
 import com.mangopay.MangoPayApi;
 import com.mangopay.core.APIs.ApiBase;
-import com.mangopay.core.APIs.declaration.ApiRefunds;
+import com.mangopay.core.APIs.ApiRefunds;
 import com.mangopay.entities.Refund;
 
 /**
